Broke apart mupr.connection into more cohesive submodules
