+++ mupt/mupr/connection/alignment.py
@@ -1,0 +1,7 @@
+'''Strategies for ensuring spatial alignment of pairs on Connectors forming a connection'''
+
+__author__ = 'Timotej Bernat'
+__email__ = 'timotej.bernat@colorado.edu'
+
+# TODO: separate impls from Connector methods (currently rigid and ballistic) and move here as strategies
+...
