+++ mupt/mupr/connection/connectors.py
@@ -1,4 +1,8 @@
-'''Abstractions of connections between two primitives'''
+'''
+Core components of connections, namely:
+* AttachmentPoints, which define geometric positions and selectivity of attachment sites
+* Connectors, which comprise 2 attachment points (an "anchor" and a "linker") and represent half of a chemical bond
+'''
 
 __author__ = 'Timotej Bernat'
 __email__ = 'timotej.bernat@colorado.edu'
@@ -8,63 +12,31 @@ LOGGER = logging.getLogger(__name__)
 
 from typing import (
     Any,
-    Callable,
     ClassVar,
-    Collection,
     Generator,
     Hashable,
     Iterable,
-    Mapping,
     Optional,
-    Protocol,
-    TypeVar,
-    TypeAlias,
     Union,
 )
 from warnings import warn
 
 from dataclasses import dataclass, field
-from enum import Enum
 from copy import deepcopy
 from itertools import product as cartesian
 
 import numpy as np
 from scipy.spatial.transform import Rotation, RigidTransform
 
-from .canonicalize import lex_order_multiset_str
-from ..chemistry.core import BondType, BOND_ORDER
-from ..geometry.arraytypes import Shape, Vector3, Array3x3, as_n_vector, compare_optional_positions
-from ..geometry.measure import within_ball
-from ..geometry.coordinates.basis import is_orthonormal
-from ..geometry.transforms.linear import rejector
-from ..geometry.transforms.rigid.rotations import alignment_rotation
-from ..geometry.transforms.rigid.application import RigidlyTransformable
-
-
-# Label typehints
-type AttachmentLabel = Hashable  # TODO: narrow down this type as use cases become clearer
-type ConnectorLabel = Hashable
-ConnectorHandle = tuple[ConnectorLabel, int]
-
-ConnectorAddress = TypeVar('ConnectorAddress', bound=int)
-
-
-# Custom Exceptions
-class ConnectionError(Exception):
-    '''Raised when Connector-related errors as encountered'''
-    pass
-
-class IncompatibleConnectorError(ConnectionError):
-    '''Raised when attempting to connect two Connectors which are, for whatever reason, incompatible'''
-    pass
-
-class MissingConnectorError(ConnectionError):
-    '''Raised when a required Connector is missing'''
-    pass
-
-class UnboundConnectorError(ConnectionError):
-    '''Raised when a pair of Connectors are unexpectedly not bound to one another'''
-    pass
+from .types import AttachmentLabel, ConnectorLabel, ConnectorHandle
+from ..canonicalize import lex_order_multiset_str
+from ...chemistry.core import BondType, BOND_ORDER
+from ...geometry.arraytypes import Vector3, Array3x3, as_n_vector, compare_optional_positions
+from ...geometry.measure import within_ball
+from ...geometry.coordinates.basis import is_orthonormal
+from ...geometry.transforms.linear import rejector
+from ...geometry.transforms.rigid.rotations import alignment_rotation
+from ...geometry.transforms.rigid.application import RigidlyTransformable
 
 
 # DEV: would love to make this frozen, but that breaks the RigidlyTansformable mechanism under-the-hood,
@@ -121,6 +93,14 @@ class Connector(RigidlyTransformable):
         self.metadata = metadata or dict()
     
         self._tangent_position = None # DEV: no call to setter; must be assigned via protected tangent_vector property
+
+    @property
+    def bond_order(self) -> float:
+        '''
+        A numerical bond order corresponding to the type of bond this Connector is associated with
+        E.g. UNASSIGNED = 0.0, AROMATIC = 1.5, DOUBLE = 2.0, etc.
+        '''
+        return BOND_ORDER.get(self.bondtype, 0.0)
 
     # Geometric properties
     ## DEV: implemented vector properties (e.g. bond/tangent/normal) by tracking endpoint positions under the hood to get them to
@@ -623,28 +603,6 @@ class Connector(RigidlyTransformable):
         
         return counterpart
 
-## Selection between pairs of Connectors (useful, for example, for resolution-shift operations)
-ConnectorSelector : TypeAlias = Callable[[Connector, Connector], Connector]
-
-def select_first(connector1 : Connector, connector2 : Connector) -> Connector:
-    '''Select the first of a pair of Connectors'''
-    return connector1
-
-def select_second(connector1 : Connector, connector2 : Connector) -> Connector:
-    '''Select the second of a pair of Connectors'''
-    return connector2
-
-def make_second_resemble_first(connector1 : Connector, connector2 : Connector) -> Connector:
-    '''Select the first of a pair of Connectors, but merge their linkables'''
-    new_connector = connector2.copy()
-    new_connector.anchor.attachables.update(connector1.anchor.attachables)
-    new_connector.linker.attachables.update(connector1.linker.attachables)
-    
-    return new_connector
-
-# DEV: provide implementations which make some attempt to reconcile spatial info attache to respective Connectors
-...
-
 # Canonicalization
 def canonical_form_connectors(connectors: Iterable[Connector], separator : str=':', joiner : str='-') -> str:
     '''A hashable string representing a collection of Connectors in canonical form'''
@@ -654,36 +612,3 @@ def canonical_form_connectors(connectors: Iterable[Connector], separator : str='
         separator=separator,
         joiner=joiner,
     )
-
-class ManagesConnectors(Protocol):
-    '''Interface for objects which manage Connectors and pairs of Connectors ("connections")'''
-    connectors : Collection[Connector]
-    connectors_by_address : Mapping[ConnectorAddress, Connector]
-    
-    def connector(self, conn_addr : ConnectorAddress) -> Connector:
-        ...
-
-    @property
-    def connectors_free(self) -> Collection[Connector]:
-        '''Connectors which are currently unbound'''
-        ...
-
-    @property
-    def connectors_bound(self) -> Collection[Connector]:
-        '''Connectors which have a neighbor'''
-        ...
-
-    # default implementations, for when explicitly inherited
-    @property
-    def functionality(self) -> int:
-        return len(self.connectors_free)
-    
-    @property
-    def valence(self) -> int: # DEV: well-defined from more than just atomic primitives since Connectors store BondType info
-        '''Electronic valence of the Primitive, i.e. the total bond order of all external-facing Connectors on this Primitive'''
-        total_bond_order : float = sum(
-            BOND_ORDER.get(conn.bondtype, 0.0)
-                for conn in self.connectors
-        )
-        return round(total_bond_order)
-    chemical_valence = electronic_valence = valence # aliases for convenience
